mavlink: move GIMBAL_MANAGER_STATUS to separate stream header
